Automatic Sync from SCM-svn to GitHub syncing the following commits: TomasTrnka: Retry workdir deletion on Windows if it fails due to open files SO-- SCMSUITE-5056
This is not the proper fix and it's unclear if it helps often enough. It
would be better to track down all the child processes and kill them
somehow.
TomasTrnka: Make AMSWorker work on Windows SO-- SCMSUITE-5056
TomasTrnka: Fix waiting for the worker to exit SO-- SCMSUITE-5056

Previously, we wouldn't wait at all if the Exit() call failed (such as
when the worker initiated the shutdown). We would then kill() our
immediate child ("sh"), leaving the actual "ams" process running in the
background while removing its files underneath it.
TomasTrnka: Properly mark static methods in AMSWorkerPool SO-- SCMSUITE-5056