Speed up unit lookups using a prebuilt dict
The previous method of iterating over all dicts in every call became
unbearably slow as new units are added, especially due to millions of
redundant calls to str.lower(). Precomputing the mapping of lowercased
unit names to possible quantities reduces that to just two lookups in
the same dict, speeding the process up by a factor of 50.